refactor : 추상화에 따른 코드 로직 이전
2022. 02. 13
AuthenticationInterceptor
- Session, Token 공통로직 이전

TokenAuthenticationInterceptor2, SeesionAuthenticationInterceptor2
- 코드 이전에 따른 간소화

Resolver : loop-study
See also : heesunjeong
