+++ src/main/java/nextstep/auth/authentication/AuthenticationInterceptor.java
@@ -1,6 +1,7 @@
 package nextstep.auth.authentication;
 
 import nextstep.auth.context.Authentication;
+import nextstep.member.domain.LoginMember;
 import org.springframework.web.servlet.HandlerInterceptor;
 
 import javax.servlet.http.HttpServletRequest;
@@ -9,10 +10,39 @@ import java.io.IOException;
 
 public abstract class AuthenticationInterceptor implements HandlerInterceptor {
     private AuthenticationConverter converter;
+    private UserDetailService userDetailService;
 
-    public AuthenticationInterceptor(AuthenticationConverter authenticationConverter) {
-        this.converter = authenticationConverter;
+    public AuthenticationInterceptor(AuthenticationConverter converter, UserDetailService userDetailService) {
+        this.converter = converter;
+        this.userDetailService = userDetailService;
+    }
+
+    @Override
+    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws IOException {
+        AuthenticationToken authenticationToken = converter.convert(request);
+        Authentication authentication = authenticate(authenticationToken);
+
+        afterAuthentication(request, response, authentication);
+        return false;
     }
 
     public abstract void afterAuthentication(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException;
+
+    public Authentication authenticate(AuthenticationToken token) {
+        String principal = token.getPrincipal();
+        LoginMember userDetails = userDetailService.loadUserByUsername(principal);
+        checkAuthentication(userDetails, token);
+
+        return new Authentication(userDetails);
+    }
+
+    private void checkAuthentication(LoginMember userDetails, AuthenticationToken token) {
+        if (userDetails == null) {
+            throw new AuthenticationException();
+        }
+
+        if (!userDetails.checkPassword(token.getCredentials())) {
+            throw new AuthenticationException();
+        }
+    }
 }
