+++ src/main/java/nextstep/auth/authentication/SessionAuthenticationInterceptor2.java
@@ -14,20 +14,8 @@ import java.util.Map;
 import static nextstep.auth.context.SecurityContextHolder.SPRING_SECURITY_CONTEXT_KEY;
 
 public class SessionAuthenticationInterceptor2 extends AuthenticationInterceptor {
-    private UserDetailService userDetailsService;
-
     public SessionAuthenticationInterceptor2(UserDetailService userDetailsService, AuthenticationConverter authenticationConverter) {
-        super(authenticationConverter);
-        this.userDetailsService = userDetailsService;
-    }
-
-    @Override
-    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) {
-        AuthenticationToken token = convert(request);
-        Authentication authentication = authenticate(token);
-
-        afterAuthentication(request, response, authentication);
-        return false;
+        super(authenticationConverter, userDetailsService);
     }
 
     @Override
@@ -35,23 +23,5 @@ public class SessionAuthenticationInterceptor2 extends AuthenticationInterceptor
         HttpSession httpSession = request.getSession();
         httpSession.setAttribute(SPRING_SECURITY_CONTEXT_KEY, new SecurityContext(authentication));
         response.setStatus(HttpServletResponse.SC_OK);
-    }
-
-    public Authentication authenticate(AuthenticationToken token) {
-        String principal = token.getPrincipal();
-        LoginMember userDetails = userDetailsService.loadUserByUsername(principal);
-        checkAuthentication(userDetails, token);
-
-        return new Authentication(userDetails);
-    }
-
-    private void checkAuthentication(LoginMember userDetails, AuthenticationToken token) {
-        if (userDetails == null) {
-            throw new AuthenticationException();
-        }
-
-        if (!userDetails.checkPassword(token.getCredentials())) {
-            throw new AuthenticationException();
-        }
     }
 }
