+++ src/main/java/nextstep/auth/authentication/TokenAuthenticationInterceptor2.java
@@ -14,25 +14,13 @@ import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
 public class TokenAuthenticationInterceptor2 extends AuthenticationInterceptor {
-
-    private UserDetailService customUserDetailsService;
     private JwtTokenProvider jwtTokenProvider;
     private ObjectMapperBean objectMapper;
 
     public TokenAuthenticationInterceptor2(UserDetailService customUserDetailsService, JwtTokenProvider jwtTokenProvider, ObjectMapperBean objectMapper, AuthenticationConverter authenticationConverter) {
-        super(authenticationConverter);
-        this.customUserDetailsService = customUserDetailsService;
+        super(authenticationConverter, customUserDetailsService);
         this.jwtTokenProvider = jwtTokenProvider;
         this.objectMapper = objectMapper;
-    }
-
-    @Override
-    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws IOException {
-        AuthenticationToken authenticationToken = convert(request);
-        Authentication authentication = authenticate(authenticationToken);
-
-        afterAuthentication(request, response, authentication);
-        return false;
     }
 
     @Override
@@ -45,23 +33,5 @@ public class TokenAuthenticationInterceptor2 extends AuthenticationInterceptor {
         response.setStatus(HttpServletResponse.SC_OK);
         response.setContentType(MediaType.APPLICATION_JSON_VALUE);
         response.getOutputStream().print(responseToClient);
-    }
-
-    public Authentication authenticate(AuthenticationToken authenticationToken) {
-        String principal = authenticationToken.getPrincipal();
-        LoginMember userDetails = customUserDetailsService.loadUserByUsername(principal);
-        checkAuthentication(userDetails, authenticationToken);
-
-        return new Authentication(userDetails);
-    }
-
-    private void checkAuthentication(LoginMember userDetails, AuthenticationToken token) {
-        if (userDetails == null) {
-            throw new AuthenticationException();
-        }
-
-        if (!userDetails.checkPassword(token.getCredentials())) {
-            throw new AuthenticationException();
-        }
     }
 }
